Fixed the mouse movement code(in-game works now as well), where if x was 0 and y was 0 in the input code it would not queue a event, should have been a AND not a OR :/.
Also made angles float from short todo fix that back.

+++ cgame/cgame_main.cs
@@ -273,9 +273,9 @@ namespace cgame
         //
         // HandleMouseEvent
         //
-        public override void HandleMouseEvent(int x, int y)
+        public override void HandleMouseEvent(float x, float y)
         {
-            
+
         }
 
         //
@@ -355,7 +355,7 @@ namespace cgame
             // Send the command to the server
             Engine.net.SendReliablePacketToAddress(idNetSource.NS_SERVER, Engine.net.GetLoopBackAddress(), ref msg);
 
-            // Dispose of the msg.
+            // Dispose of the msg
             msg.Dispose();
         }
